Very small changes, just changing 4 lines of codes.

+++ h2gis-functions/src/test/java/org/h2gis/functions/io/geojson/GeojsonImportExportTest.java
@@ -590,8 +590,8 @@ public class GeojsonImportExportTest {
         assertTrue(res.getString(10).equals("2017-01-19T18:29:27+01:00"));
         assertTrue(res.getBigDecimal(11).toString().equals("1484846967000"));
         assertTrue(res.getString(12).equals("{}"));
-        Object[] tinyArray = {(Integer) 13, "string", "{}"};
-        Object[] expectedResult = {(Integer) 49, (Double) 40.0, "{}", "string", tinyArray};
+        Object[] tinyArray = {13, "string", "{}"};
+        Object[] expectedResult = {49, 40.0, "{}", "string", tinyArray};
         Object[] result = (Object[]) res.getObject(13);
         assertArrayEquals(expectedResult, result);
         expectedResult = new Object[]{58, 47, 58, 57, 58, 49, 58, 51, 58, 58, 49, 57, 58, 58, 49, 58, 57, 56, 57, 58, 59, 58, 57, 58, 49, 47, 48, 57, 48, 58, 57, 57, 51, 56, 52, 57, 51, 57, 49, 58, 55, 58, 50, 48, 48, 52, 56, 57, 48, 58, 52, 48, 53, 50, 57, 54, 57, 47, 58, 57, 54, 54, 53, 56, 57, 55, 58, 58, 57, 58, 57, 57};
@@ -624,8 +624,8 @@ public class GeojsonImportExportTest {
         assertTrue(res.getString(10).equals("2017-01-19T18:29:27+01:00"));
         assertTrue(res.getBigDecimal(11).toString().equals("1484846967000"));
         assertTrue(res.getString(12).equals("{}"));
-        Object[] tinyArray = {(Integer) 13, "string", "{}"};
-        Object[] expectedResult = {(Integer) 49, (Double) 40.0, "{}", "string", tinyArray};
+        Object[] tinyArray = {13, "string", "{}"};
+        Object[] expectedResult = {49, 40.0, "{}", "string", tinyArray};
         Object[] result = (Object[]) res.getObject(13);
         assertArrayEquals(expectedResult, result);
         expectedResult = new Object[]{58, 47, 58, 57, 58, 49, 58, 51, 58, 58, 49, 57, 58, 58, 49, 58, 57, 56, 57, 58, 59, 58, 57, 58, 49, 47, 48, 57, 48, 58, 57, 57, 51, 56, 52, 57, 51, 57, 49, 58, 55, 58, 50, 48, 48, 52, 56, 57, 48, 58, 52, 48, 53, 50, 57, 54, 57, 47, 58, 57, 54, 54, 53, 56, 57, 55, 58, 58, 57, 58, 57, 57};
